requirements.txt
# Web API
fastapi==0.111.0
uvicorn[standard]==0.30.1
orjson==3.10.3
pydantic==2.7.1
starlette==0.37.2

# Core scientific
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.11.4
tqdm==4.66.4

# IR: lexical + dense + rerank
rank-bm25==0.2.2
faiss-cpu==1.8.0
sentence-transformers==2.6.1
transformers==4.41.2
tokenizers==0.19.1
huggingface-hub==0.23.4
torch==2.2.2
safetensors==0.4.3

# Content extraction
trafilatura==1.6.3
lxml==5.2.2
beautifulsoup4==4.12.3
httpx==0.27.0

# File formats
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.1
loguru==0.7.2

# Testing / lint
pyarrow==16.1.0
pytest==8.2.1
pytest-asyncio==0.23.6
httpx[http2]==0.27.0
ruff==0.4.8
black==24.4.2
mypy==1.10.0
